market compo up and running
