+++ src/stores/userStore.js
@@ -19,43 +19,42 @@ export class UserStore {
       "available cash": 6000,
       "average return": 7.8,
       "next payment": "15-02-19"
-
     }
-  //   this.openLoans = [{
-  //     "username":  "shooobert",
-  //     "period": 2400,
-  //     "purpose": "coding bootcamp",
-  //     "amount": 500,
-  //     "interest": 500,
-  //     "remaining amount": 4300,
-  //     "status": "OK",
-  //     "issuance date": "15-09-19",
-  //     "next payment": "15-01-20"
-  //   },
-  //   {
-  //     "username":  "dudi",
-  //     "period": 2400,
-  //     "purpose": "coding bootcamp",
-  //     "amount": 500,
-  //     "interest": 500,
-  //     "remaining amount": 4300,
-  //     "status": "OK",
-  //     "issuance date": "15-09-19",
-  //     "next payment": "15-01-20"
-  //   },
-  //   {
-  //     "username":  "rachel",
-  //     "period": 2400,
-  //     "purpose": "coding bootcamp",
-  //     "amount": 500,
-  //     "interest": 500,
-  //     "remaining amount": 4300,
-  //     "status": "OK",
-  //     "issuance date": "15-09-19",
-  //     "next payment": "15-01-20"
-  //   }
+    this.openLoans = [{
+      "username":  "shooobert",
+      "period": 2400,
+      "purpose": "coding bootcamp",
+      "amount": 500,
+      "interest": 500,
+      "remaining amount": 4300,
+      "status": "OK",
+      "issuance date": "15-09-19",
+      "next payment": "15-01-20"
+    },
+    {
+      "username":  "dudi",
+      "period": 2400,
+      "purpose": "coding bootcamp",
+      "amount": 500,
+      "interest": 500,
+      "remaining amount": 4300,
+      "status": "OK",
+      "issuance date": "15-09-19",
+      "next payment": "15-01-20"
+    },
+    {
+      "username":  "rachel",
+      "period": 2400,
+      "purpose": "coding bootcamp",
+      "amount": 500,
+      "interest": 500,
+      "remaining amount": 4300,
+      "status": "OK",
+      "issuance date": "15-09-19",
+      "next payment": "15-01-20"
+    }
   
-  // ]
+  ]
 
     console.log(this.user)
     // let data = await axios.get(dataRoute);
